VerifyBot/VerifyBot/web/requirements.txt
flask
fastapi
uvicorn
jinja2
geoip2
python-multipart
requests
gspread
python-dotenv
google-auth==2.28.1
gspread_formatting